+++ src/Farmer/Builders/Builders.CommunicationServices.fs
@@ -2,25 +2,30 @@
 module Farmer.Builders.CommunicationServices
 
 open Farmer
-open Farmer.Arm.CommunicationServices
+open Farmer.Arm.Communication
 
 type CommunicationServices =
-    /// Gets an ARM Expression key for any Bing Search instance.
+    /// Gets an ARM Expression key for any Communication Services instance.
     static member getKey (resourceId: ResourceId) =
-        ArmExpression.create($"listKeys({resourceId.ArmExpression.Value}, '{resource.ApiVersion}').key1", resourceId)
-    static member getKey (name: ResourceName) = CommunicationServices.getKey (resource.resourceId name)
+        ArmExpression.create($"listKeys({resourceId.ArmExpression.Value}, '{communicationServices.ApiVersion}').primaryKey", resourceId)
+    static member getKey (name: ResourceName) = CommunicationServices.getKey (communicationServices.resourceId name)
+
+    static member getConnectionString (resourceId: ResourceId) =
+        ArmExpression.create($"listKeys({resourceId.ArmExpression.Value}, '{communicationServices.ApiVersion}').primaryConnectionString", resourceId)
+    static member getConnectionString (name: ResourceName) = CommunicationServices.getConnectionString (communicationServices.resourceId name)
 
 type CommunicationServicesConfig =
     { Name: ResourceName
       Tags: Map<string,string>
       DataLocation: DataLocation }
-    /// Gets an ARM expression to the key of this Bing Search instance.
-    member this.Key = CommunicationServices.getKey (resource.resourceId this.Name)
+    /// Gets an ARM expression to the key of this Communication Services instance.
+    member this.Key = CommunicationServices.getKey (communicationServices.resourceId this.Name)
+    /// Gets an ARM expression to the connection string of this Communication Services instance.
+    member this.ConnectionString = CommunicationServices.getConnectionString (communicationServices.resourceId this.Name)
     interface IBuilder with
-        member this.ResourceId = resource.resourceId this.Name
-        member this.BuildResources location = [
-            { Name = this.Name
-              Location = location
+        member this.ResourceId = communicationServices.resourceId this.Name
+        member this.BuildResources _ = [
+            { CommunicationService.Name = this.Name
               Tags = this.Tags
               DataLocation = this.DataLocation }
         ]
@@ -34,7 +39,7 @@ type CommunicationServicesBuilder () =
     [<CustomOperation "name">]
     member _.Name (state: CommunicationServicesConfig, name) = { state with Name = ResourceName name }
     [<CustomOperation "data_location">]
-    member _.Sku (state: CommunicationServicesConfig, dataLocation) = { state with DataLocation = dataLocation }
+    member _.DataLocation (state: CommunicationServicesConfig, dataLocation) = { state with DataLocation = dataLocation }
     interface ITaggable<CommunicationServicesConfig> with member _.Add state tags = { state with Tags = state.Tags |> Map.merge tags }
 
-let communicationServices = CommunicationServicesBuilder()
+let communicationService = CommunicationServicesBuilder()